+++ changelog.md
@@ -1,2 +1,6 @@
 # 1.0.0
-- Edit this file to change your mod's changelog.
+- added snap option for rotation and anchor position
+- added independent (free) rotation option
+- fixed button sprite scale issue (sprite size didn't match button touch box)
+- fixed overlap of buttons with the controls after specific transforms
+- prevented zero-division crash when anchor is too close the control and you try to transform objects